% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{analize_link}
\alias{analize_link}
\title{Analize the uri  and all the links found on it}
\usage{
analize_link(uri, level, max_deepth, max_internal_links,
  domainCount = list())
}
\arguments{
\item{uri}{the uri to be analized}

\item{level}{current analisis level}

\item{max_deepth}{The maximum level to stop the analisis}

\item{max_internal_links}{The maximun number of links to be analized inside the same domain}

\item{domainCount}{How many times a domain has been analized}
}
\value{
Data frame with the link information
}
\description{
Analize the uri  and all the links found on it
}
