% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeDomains.R
\name{analizy_domains}
\alias{analizy_domains}
\title{Verify the domains analized and verifies if contains malware}
\usage{
analizy_domains(df_domains, malware_domain_list)
}
\value{
dataframe including if contains malware information
}
\description{
Verify the domains analized and verifies if contains malware
}
