% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Bar_plot_semester_Cerber}
\alias{Bar_plot_semester_Cerber}
\title{Plot Ceber occurrences (per semester)}
\usage{
Bar_plot_semester_Cerber(x)
}
\value{
Bar plot showing the Ceber occurrences (per semester)
}
\description{
This function plots, from the dataframe calculated by the SemesterColumnsMalwareTotal function, the occurrences of Ceber per semester (years [2015..2108])
}
