% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Bar_plot_semester_Locky}
\alias{Bar_plot_semester_Locky}
\title{Plot Locky occurrences (per semester)}
\usage{
Bar_plot_semester_Locky(x)
}
\value{
Bar plot showing the Locky occurrences (per semester)
}
\description{
This function plots, from the dataframe calculated by the SemesterColumnsMalwareTotal function, the occurrences of Locky per semester (years [2015..2108])
}
