% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Bar_plot_semester_Tesla}
\alias{Bar_plot_semester_Tesla}
\title{Plot TeslaCrypt occurrences (per semester)}
\usage{
Bar_plot_semester_Tesla(x)
}
\value{
Bar plot showing the TeslaCrypt occurrences (per semester)
}
\description{
This function plots, from the dataframe calculated by the SemesterColumnsMalwareTotal function, the occurrences of TeslaCrypt per semester (years [2015..2108])
}
