% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Bar_plot_year}
\alias{Bar_plot_year}
\title{Plot all the malware occurrences per year}
\usage{
Bar_plot_year(x, year)
}
\arguments{
\item{df}{year}
}
\value{
Bar plot showing the TeslaCrypt occurrences (per semester)
}
\description{
This function plots the occurrences of all the malware listed in the csv file during
 one year (shown by semester)
}
