% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Bar_plot_year_scale}
\alias{Bar_plot_year_scale}
\title{Plot all the malware occurrences per year (scaled)}
\usage{
Bar_plot_year_scale(x, year, yMax)
}
\arguments{
\item{df}{year yMax}
}
\value{
Bar plot showing the TeslaCrypt occurrences (per semester)
}
\description{
This function plots the occurrences of the malware listed in the csv file during one year
(shown by semester) whose number of occurrences are under the yMax value
}
