% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpositionCrawlingData.R
\name{build_domain_graph}
\alias{build_domain_graph}
\title{Build a graph from the dataframe links}
\usage{
build_domain_graph(domainLinksDf, max_number_domains = 10)
}
\arguments{
\item{domainLinksDf}{Links of the domain analized}

\item{max_number_domains}{maximun number of domain to display to avoid many noise on the graph}
}
\description{
Build a graph from the dataframe links
}
