% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpositionMalwareData.R
\name{build_malware_graph}
\alias{build_malware_graph}
\title{Build a graph including only paths to the malware}
\usage{
build_malware_graph(domainLinksDf, max_number_domains = 10)
}
\arguments{
\item{domainLinksDf}{Links of the domain analized}

\item{max_number_domains}{maximun number of domain to display to avoid many noise on the graph}
}
\description{
Green edge are safe domains
Red edges are dangerous domains
}
