% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{changeColumnName}
\alias{changeColumnName}
\title{Change column name}
\usage{
changeColumnName(df, previous, new)
}
\arguments{
\item{df}{previous new}
}
\value{
Dataframe with the new column name
}
\description{
This function replaces in the dataframe (df) the name of the column (previous) with a new column name (new)
}
