% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{changeDateFormat}
\alias{changeDateFormat}
\title{Change date format}
\usage{
changeDateFormat(df, previous, new)
}
\arguments{
\item{df}{previous new}
}
\value{
Data frame with the new date format in column DateHour
}
\description{
This function changes the date format for the column "DateHour" of the "df" dataframe. The previous date format is "previous"; the new one is "new".
For exemple:
'%Y-%m-%d %H:%M:%S' to '%Y-%m-%d'
}
