% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{clean_and_transform_links_dataframe}
\alias{clean_and_transform_links_dataframe}
\title{Clean the links data frame removing NA values and adding more information to the data from the link}
\usage{
clean_and_transform_links_dataframe(uri, linksDataFrame)
}
\arguments{
\item{uri}{uri where the links was found}

\item{linksDataFrame}{Original data frames with the links}
}
\value{
The clean and extended data frame
}
\description{
The added information is the level, scheme, domain, path and parameters
}
