% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{CountryMalwareTotal}
\alias{CountryMalwareTotal}
\title{Dataframe by country}
\usage{
CountryMalwareTotal(x, country)
}
\arguments{
\item{df}{year}
}
\value{
Dataframe
}
\description{
This function arranges a dataframe (malware by country)
}
