% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{DeleteColumns}
\alias{DeleteColumns}
\title{Delete columns}
\usage{
DeleteColumns(df)
}
\value{
Data frame without the deleted columns
}
\description{
This function changes delete some columns of df dataframe:
(Host, URL,Registrar, IPaddress.es, ASN.s)
}
