% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{DeleteColumnsCountry}
\alias{DeleteColumnsCountry}
\title{Delete columns country}
\usage{
DeleteColumnsCountry(df)
}
\value{
Data frame without the deleted columns
}
\description{
This function deletes some columns of df dataframe:
(Host, URL,Registrar, IPaddress.es)
}
