% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{get_links}
\alias{get_links}
\title{Extract the links from the current page and creates a data frame from the links}
\usage{
get_links(uri, page, level)
}
\arguments{
\item{uri}{uri where the links was found}

\item{page}{the xml parsed website}

\item{level}{current analize level to be included on the data frame}
}
\value{
Data frame with the link information
}
\description{
Extract the links from the current page and creates a data frame from the links
}
