% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalwareData.R
\name{get_malware_domain_list}
\alias{get_malware_domain_list}
\title{Returns a list of malware domains}
\usage{
get_malware_domain_list()
}
\value{
malware domains
}
\description{
Returns a list of malware domains
}
