% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{load_domain}
\alias{load_domain}
\title{The function analize the given domain and domains accesible from them}
\usage{
load_domain(uri, max_internal_links = 10, max_deepth = 5)
}
\arguments{
\item{uri}{uri to analize}

\item{max_internal_links}{maximun links to follown on the same domain, this is just to avoid the crawler take many time on domains with many links}

\item{max_deepth}{maximun number to external domains to follow}
}
\description{
The function analize the given domain and domains accesible from them
}
\details{
This are the details
}
