% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{load_page}
\alias{load_page}
\title{This function download the html}
\usage{
load_page(uri)
}
\arguments{
\item{uri}{uri to parse as XML}
}
\value{
The uri parse as XML
}
\description{
This function download the html
}
\details{
The function download the html.
SSL errors are ignored using the httr configuration ssl_verifypeer
Error downloading webpages due timeouts, incorrect uris our website nor available temporary are ingored and parsed as empty websites
}
