% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Plot_bar_horizontal}
\alias{Plot_bar_horizontal}
\title{Plot all the malware occurrences per country}
\usage{
Plot_bar_horizontal(x, yg)
}
\arguments{
\item{df}{year}
}
\value{
Bar plot showing the ranswomware by country
}
\description{
This function plots the occurrences of all the malware listed in the csv file per
country
}
