% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Plot_pie_Max}
\alias{Plot_pie_Max}
\title{Percentage of occurrences (Pie Chart)}
\usage{
Plot_pie_Max(x)
}
\value{
Pie Chart showing the highest percentage of occurrences
}
\description{
This function shows in a graphical way (Pie Chart) the dataframe calculated
by the function SumColumnsMalware (percentage of malware occurrences)
}
