% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalwareData.R
\name{read_malware_data}
\alias{read_malware_data}
\title{Returns a dataframe with malware domain information}
\usage{
read_malware_data()
}
\value{
A dataframe with malware information with the following columns:
Date, Domain, domainParsed, IP, ReverseLookup, Description, ASN, Country
}
\description{
This return the malware informaiton extracted from http://www.malwaredomainlist.com
}
