% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalwareData.R
\name{read_malware_domains_from_ultimatehostblacklist}
\alias{read_malware_domains_from_ultimatehostblacklist}
\title{Returns a list of domains blocked by Ultimate.Hosts.Blacklist}
\usage{
read_malware_domains_from_ultimatehostblacklist()
}
\value{
domains blocked by Ultimate.Hosts.Blacklist
}
\description{
Returns a list of domains blocked by Ultimate.Hosts.Blacklist
}
