% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{SemesterColumnsMalwareTotal}
\alias{SemesterColumnsMalwareTotal}
\title{Malware occurrences per semester (2015-2018)}
\usage{
SemesterColumnsMalwareTotal(x)
}
\value{
Dataframe showing the occurrences of each malware (per semester)
}
\description{
This function lists the occurrences of each one of the principal malware (TeslaCrypt,
Cerber and Locky) for each semester during the years between 2015 and 2018
}
