% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{separateDate}
\alias{separateDate}
\title{Separate date columns}
\usage{
separateDate(df)
}
\value{
Dataframe with the two new columns: "Date" and "Hour"
}
\description{
This function separates the "DateHour" column of the df dataframe into new
columns "Date" and "Hour. The separator must be " "
}
