% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{SumColumnsMalware}
\alias{SumColumnsMalware}
\title{Percentage of occurrences}
\usage{
SumColumnsMalware(x)
}
\value{
Dataframe showing the highest percentage of occurrences
}
\description{
This function selects the first three malware (according to the number of occurrences)
and sums all the occurrences of the other nine malwares in a new row called "Other".
It also calculates the percentage of occurrences.
}
