% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{SumColumnsMalwareTotal}
\alias{SumColumnsMalwareTotal}
\title{Sum the ocurrences of each Ransomware in 2015-2018}
\usage{
SumColumnsMalwareTotal(x)
}
\value{
Dataframe that relates the Malware with their total occurrences
}
\description{
This function sums the total occurrences during the period between 2015 and 2018, ordering
the result in descending order (by Frequency)
}
