% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawler.R
\name{update_domain_count}
\alias{update_domain_count}
\title{Counts how many times a domain has been analize to be able to stop the internal search inside the same domain}
\usage{
update_domain_count(currentDomain, currentDomainCount)
}
\arguments{
\item{currentDomain}{the current domain analized}

\item{currentDomainCount}{the number of times the domains has been found}
}
\value{
Data frame with the link information
}
\description{
Counts how many times a domain has been analize to be able to stop the internal search inside the same domain
}
