% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalwareData.R
\name{update_malware_data}
\alias{update_malware_data}
\title{Updates malware domain information csv}
\usage{
update_malware_data()
}
\description{
This return the malware informaiton extracted from http://www.malwaredomainlist.com
}
